move update tests to separate file
